web/requirements.txt
Django==1.11.6
django-rest-auth==0.9.2
django-rest-framework==0.1.0
djangorestframework==3.7.1
pytz==2017.2
six==1.11.0
